+++ HISTORY.rst
@@ -10,6 +10,7 @@ Release History
 
 - 3.14 utc market fix
 - seconds_to_start bug when set to 0 seconds
+- limit order persistence validation (MarketOnClose when market is inplay or no inplay available)
 
 2.11.0 (2025-12-11)
 +++++++++++++++++++